+++ test/route.cpp
@@ -26,7 +26,7 @@ TEST_CASE("Route", "[route]") {
     auto out2 = capture_output<event_set<tc_event, marker_event<>>>();
     auto in = feed_input<event_set<tc_event, marker_event<>>>(
         route<event_set<tc_event>>(
-            channel_router(std::array<std::int16_t, 3>{5, -3, -32768}),
+            channel_router(std::array<std::int32_t, 3>{5, -3, -32768}),
             ref_processor(out0), ref_processor(out1), ref_processor(out2)));
     in.require_output_checked(out0);
     in.require_output_checked(out1);
